Improve scaling on large number of ranks.
Use MPI_Iallreduce on MPI3 to overlap two calls to Allreduce.
The operation can take a long time when number ranks are large, as
we are reducing an array of NTask in size.

Implement a faster gathered sorting routine for small data sets.
data will be collected to rank 0, sorted, and scattered.
Definition of small is if the total data is less than 32 * NTask in bytes.
Since we always need about this much data in histogram sort for book
keeping, this is a reasonable assumption.

Disable the internal sparse Alltoallv implementation by default. Usually
underlying MPI provides variables to control these behaviors at runtime.

Added interface to set the tuning flags.