Use a simpler overlapping.
The finer overlapping turns out to be making each of the
iterations 2x slower.

The simpler overlapping only overlaps the histograming of one
of the CLT/CLE edges. Reducing the number of iallreduce seems
to help the performance. But overall it is barely faster than
the straight two blocking allreduce call version;

It is because either:
- the histogramming is still too fast
- Cray's current nonblocking implementation doesn't actually
do anywork till wait is issued, or used an algorithm that is
very heavy on CPU.