_assign_colors had an argument, outsizes, which was not initialized
outsizes was passed down to _assign_colors without being initialized. It
was never used after that function, but within that function it could be
used to set avoid setting mycolor = -1. This didn't show up when it was
a stack-allocated VLA (perhaps it was being zerod by virtue of being on
the same page as sizes?) but when I made it malloc'd it caused a test
failure.

Remove this argument to the functions instead of just memsetting it
as it is not used and I could not figure out what it was for.

This is the root cause of the bugs where sparse data fails to sort!