Introducing mpiu.c.
This is mostly code refactor, with a possible fix to issue#10.

memory managment is now done with MPIU_SetMalloc. All routines that uses
MPIU shall use MPIU_Malloc for large memory allocations.

The idea is to share mpiu.[ch] between mpsort, bigfile, MP-Gadget and FastPM.
Will add a new repo for mpiu.[ch], and either keep duplicating the files,
or add a new library; but first, convert each users to this format.